Refresh shapes each time
that's why chrome had a phantom issue with transparency -- it's only on second run that it was causing a problem. we don't clear the scene if you don't change the file. chrome's renderer multiplies the transparent objects but ff only shows one

and fix the base url on the default STL link -- only a problem in dev (and maybe still only in chrome?)